Move menu form to separate component
