+++ src/components/features/MenuForm/MenuForm.module.scss
@@ -1,0 +1,8 @@
+.input {
+  width: 200px;
+  margin: 20px;
+}
+
+.card {
+  flex-direction: column;
+}
